refactor: split app into internal/cli + app
