+++ internal/cli/flags.go
@@ -1,4 +1,4 @@
-package app
+package cli
 
 import (
 	"errors"
