+++ internal/cli/flags_test.go
@@ -1,4 +1,4 @@
-package app
+package cli
 
 import (
 	"os"
