+++ internal/cli/parser.go
@@ -1,4 +1,4 @@
-package app
+package cli
 
 import (
 	"fmt"
